pi/requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-requests==3.2.3
adafruit-circuitpython-rgb-display==3.12.4
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
arandr==0.1.11
arrow==1.3.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.2.0
av==10.0.0
Babel==2.10.3
beautifulsoup4==4.11.2
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
comm==0.2.2
contourpy==1.2.1
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
DBussy==1.3
debugpy==1.8.1
decorator==5.1.1
dill==0.3.6
distro==1.8.0
docutils==0.19
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
Flask==2.2.2
fonttools==4.51.0
fsspec==2024.3.1
gpiozero==2.0
gradio==4.26.0
gradio_client==0.15.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==0.22.2
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.12.0
importlib_resources==6.4.0
install==1.3.5
ipykernel==6.29.4
ipython==8.23.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiwer==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libevdev==0.5
logilab-common==1.9.8
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
mypy==1.0.1
mypy-extensions==0.4.3
nest_asyncio==1.6.0
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pgzero==1.2
picamera2==0.3.17
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pyftdi==0.55.3
pygame==2.1.2
Pygments==2.17.2
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.9
python-networkmanager==2.2
python-prctl==1.8.1
pytz==2022.7.1
pyudev==0.24.0
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==26.0.0
rapidfuzz==3.8.1
referencing==0.34.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rich==13.7.1
roman==3.3
rpds-py==0.18.0
RPi.GPIO==0.7.1a4
rpi_ws281x==5.0.0
RTIMULib==7.2.1
ruff==0.3.7
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.1b0
sense-hat==2.6.0
shellingham==1.5.4
simplejpeg==1.6.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
st7735==1.0.0
stack-data==0.6.3
starlette==0.37.2
sysv-ipc==1.1.0
tabulate==0.8.9
thonny==4.1.4
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
twython==3.8.2
typer==0.12.3
types-aiofiles==22.1
types-annoy==1.17
types-appdirs==1.4
types-aws-xray-sdk==2.10
types-babel==2.11
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.11
types-bleach==5.0
types-boto==2.49
types-braintree==4.17
types-cachetools==5.2
types-caldav==0.10
types-certifi==2021.10.8
types-cffi==1.15
types-chardet==5.0
types-chevron==0.14
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.7
types-contextvars==2.4
types-croniter==1.3
types-cryptography==3.3
types-D3DShot==0.1
types-dateparser==1.1
types-DateTimeRange==1.2
types-decorator==5.1
types-Deprecated==1.2
types-dj-database-url==1.0
types-docopt==0.6
types-docutils==0.19
types-editdistance==0.6
types-emoji==2.1
types-entrypoints==0.4
types-first==2.0
types-flake8-2020==1.7
types-flake8-bugbear==22.10.27
types-flake8-builtins==2.0
types-flake8-docstrings==1.6
types-flake8-plugin-utils==1.3
types-flake8-rst-docstrings==0.2
types-flake8-simplify==0.19
types-flake8-typing-imports==1.14
types-Flask-Cors==3.0
types-Flask-SQLAlchemy==2.5
types-fpdf2==2.5
types-gdb==12.1
types-google-cloud-ndb==1.11
types-hdbcli==2.14
types-html5lib==1.1
types-httplib2==0.21
types-humanfriendly==10.0
types-invoke==1.7
types-JACK-Client==0.5
types-jmespath==1.0
types-jsonschema==4.17
types-keyboard==0.13
types-ldap3==2.9
types-Markdown==3.4
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.1
types-oauthlib==3.2
types-openpyxl==3.0
types-opentracing==2.4
types-paho-mqtt==1.6
types-paramiko==2.11
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.15
types-pep8-naming==0.13
types-Pillow==9.3
types-playsound==1.3
types-polib==1.1
types-prettytable==3.4
types-protobuf==3.20
types-psutil==5.9
types-psycopg2==2.9
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycurl==7.45
types-pyfarmhash==0.3
types-pyflakes==2.5
types-Pygments==2.13
types-pyinstaller==5.6
types-PyMySQL==1.0
types-pynput==1.7
types-pyOpenSSL==22.1
types-pyRFC3339==1.1
types-PyScreeze==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==2.6
types-python-dateutil==2.9.0.20240316
types-python-gflags==3.1
types-python-jose==3.3
types-python-nmap==0.7
types-python-slugify==6.1
types-pytz==2022.6
types-pyvmomi==7.0
types-pywin32==304
types-PyYAML==6.0
types-redis==4.3
types-regex==2022.10.31
types-requests==2.28
types-retry==0.9
types-Send2Trash==1.8
types-setuptools==65.5
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-SQLAlchemy==1.4.43
types-stdlib-list==0.8
types-stripe==3.5
types-tabulate==0.9
types-termcolor==1.1
types-toml==0.10
types-toposort==1.7
types-tqdm==4.64
types-tree-sitter==0.20
types-tree-sitter-languages==1.5
types-ttkthemes==3.2
types-typed-ast==1.5
types-tzlocal==4.2
types-ujson==5.5
types-urllib3==1.26
types-vobject==0.9
types-waitress==2.1
types-whatthepatch==1.0
types-xmltodict==0.13
types-xxhash==3.0
types-zxcvbn==4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.12
uvicorn==0.29.0
v4l2-python3==0.3.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.2
wrapt==1.14.1
zeroconf==0.47.3
zipp==1.0.0
